requirements.txt
flet
Pillow
rembg
onnxruntime